Re-export ShutdownError from common location
