+++ opentelemetry-sdk/src/error.rs
@@ -1,6 +1,4 @@
 //! Wrapper for error from trace, logs and metrics part of open telemetry.
-use std::sync::PoisonError;
-
 #[cfg(feature = "logs")]
 use crate::logs::LogError;
 #[cfg(feature = "metrics")]
@@ -8,8 +6,13 @@ use crate::metrics::MetricError;
 use opentelemetry::propagation::PropagationError;
 #[cfg(feature = "trace")]
 use opentelemetry::trace::TraceError;
+use std::sync::PoisonError;
+use std::time::Duration;
+use thiserror::Error;
 
-/// Wrapper for error from both tracing and metrics part of open telemetry.
+/// Wrapper for error from both tracing and metrics part of open telemetry. This
+/// gives us a common error type where we _need_ to return errors that may come
+/// from various components.
 #[derive(thiserror::Error, Debug)]
 #[non_exhaustive]
 pub enum Error {
@@ -34,6 +37,10 @@ pub enum Error {
     /// Error happens when injecting and extracting information using propagators.
     Propagation(#[from] PropagationError),
 
+    /// Failed to shutdown an exporter
+    #[error(transparent)]
+    Shutdown(#[from] ShutdownError),
+
     #[error("{0}")]
     /// Other types of failures not covered by the variants above.
     Other(String),
@@ -42,5 +49,30 @@ pub enum Error {
 impl<T> From<PoisonError<T>> for Error {
     fn from(err: PoisonError<T>) -> Self {
         Error::Other(err.to_string())
+    }
+}
+
+/// Errors returned by shutdown operations in the Export API.
+#[derive(Error, Debug)]
+#[non_exhaustive]
+pub enum ShutdownError {
+    /// Shutdown timed out before completing.
+    #[error("Shutdown timed out after {0:?}")]
+    Timeout(Duration),
+
+    /// The export client failed while holding the client lock. It is not
+    /// possible to complete the shutdown and a retry will not help.
+    /// This is something that should not happen and should likely emit some diagnostic.
+    #[error("export client failed while holding lock; cannot retry.")]
+    ClientFailed(String),
+
+    /// An unexpected error occurred during shutdown.
+    #[error(transparent)]
+    Other(#[from] Box<dyn std::error::Error + Send + Sync + 'static>),
+}
+
+impl<T> From<PoisonError<T>> for ShutdownError {
+    fn from(err: PoisonError<T>) -> Self {
+        ShutdownError::ClientFailed(format!("Mutex poisoned during shutdown: {}", err))
     }
 }
